+++ src/main/java/com/realworld/project/article/api/dto/RequestUpdateArticle.java
@@ -1,9 +1,7 @@
-package com.realworld.project.article.dto;
+package com.realworld.project.article.api.dto;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Getter;
-
-import javax.validation.constraints.NotEmpty;
 
 @Getter
 @JsonRootName("article")
