+++ src/main/java/com/realworld/project/article/api/dto/ResponseTag.java
@@ -1,6 +1,5 @@
-package com.realworld.project.article.dto;
+package com.realworld.project.article.api.dto;
 
-import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Getter;
 
 import java.util.List;
