+++ src/main/java/com/realworld/project/user/api/FollowApi.java
@@ -1,6 +1,6 @@
 package com.realworld.project.user.api;
 
-import com.realworld.project.user.dto.ResponseProfile;
+import com.realworld.project.user.api.dto.ResponseProfile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
