+++ src/main/java/com/realworld/project/user/api/dto/RequestLoginUser.java
@@ -1,4 +1,4 @@
-package com.realworld.project.user.dto;
+package com.realworld.project.user.api.dto;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Getter;
@@ -7,10 +7,7 @@ import javax.validation.constraints.NotEmpty;
 
 @Getter
 @JsonRootName("user")
-public class RequestRegisterUser {
-    @NotEmpty
-    private String username;
-
+public class RequestLoginUser {
     @NotEmpty
     private String email;
