+++ src/main/java/com/realworld/project/user/api/dto/RequestRegisterUser.java
@@ -1,4 +1,4 @@
-package com.realworld.project.user.dto;
+package com.realworld.project.user.api.dto;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Getter;
@@ -7,7 +7,10 @@ import javax.validation.constraints.NotEmpty;
 
 @Getter
 @JsonRootName("user")
-public class RequestLoginUser {
+public class RequestRegisterUser {
+    @NotEmpty
+    private String username;
+
     @NotEmpty
     private String email;
